requirements.txt
requests==2.28.0
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0